Remade the program
I redesigned the program as remove the need to re-distribute the mods via a google-drive link. Upon consideration, I realized that was a terrible idea. I have now remade the program to instead use a .txt with the official direct download links from the official mod providers. Apologies for the oversight and enjoy!